refactor reminders plan limit handling
